refactor: align folder structure with Next.js App Router conventions
Move profile component from components/about/ to components/sections/, move cosmos-layout.module.css from app/ to components/cosmos/, and remove empty lib/cosmos/ directory.
