+++ src/app/o-mnie/page.tsx
@@ -1,7 +1,7 @@
 import { SectionLabel } from '@/components/cosmos/section-label';
 import { aboutContent } from '@/lib/content/about';
 import { ScrollReveal } from '@/components/cosmos/scroll-reveal';
-import { AboutProfile } from '@/components/about/profile';
+import { AboutProfile } from '@/components/sections/profile';
 import { pageMetadata } from '@/lib/seo';
 import styles from './page.module.css';
